+++ app/src/core/operators/recommendations/index.ts
@@ -1,0 +1,3 @@
+export * from './recommendationEngine'
+export * from './recommendationStore'
+export * from './recommendationTypes'